Improvements to fit quality plots
The fit quality plots have been changed to show the average
absolute percent error, which is a less confusing metric than
the number of points with an error over a certain percent. I also
switched to limiting the number of test points such that each
dimension is plotted with the same number of test points.